multilingo-server/requirements.txt
fastapi
uvicorn
websockets
starlette
pydub
numpy
soundfile
openai-whisper
langdetect
deep-translator
sqlalchemy
tortoise-orm
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
fastapi-users 